+++ src/backend/src/modules/puterfs/lib/PuterFSProvider.js
@@ -402,74 +402,8 @@ class PuterFSProvider extends putility.AdvancedBase {
      * @returns {Promise<FSNode>}
      */
     async mkdir ({ context, parent, name, immutable }) {
-        const { actor, thumbnail } = context.values;
-
-        const svc_fslock = this.#services.get('fslock');
-        const lock_handle = await svc_fslock.lock_child(await parent.get('path'),
-                        name,
-                        MODE_WRITE);
-
-        try {
-            const ts = Math.round(Date.now() / 1000);
-            const uid = uuidv4();
-            const resourceService = this.#services.get('resourceService');
-            const svc_fsEntry = this.#services.get('fsEntryService');
-            const svc_event = this.#services.get('event');
-            const fs = this.#services.get('filesystem');
-
-            const existing = await fs.node(new NodeChildSelector(parent.selector, name));
-
-            if ( await existing.exists() ) {
-                throw APIError.create('item_with_same_name_exists', null, {
-                    entry_name: name,
-                });
-            }
-
-            const svc_acl = this.#services.get('acl');
-            if ( ! await parent.exists() ) {
-                throw APIError.create('subject_does_not_exist');
-            }
-            if ( ! await svc_acl.check(actor, parent, 'write') ) {
-                throw await svc_acl.get_safe_acl_error(actor, parent, 'write');
-            }
-
-            resourceService.register({
-                uid,
-                status: RESOURCE_STATUS_PENDING_CREATE,
-            });
-
-            const raw_fsentry = {
-                is_dir: 1,
-                uuid: uid,
-                parent_uid: await parent.get('uid'),
-                path: path.join(await parent.get('path'), name),
-                user_id: actor.type.user.id,
-                name,
-                created: ts,
-                accessed: ts,
-                modified: ts,
-                immutable: immutable ?? false,
-                ...(thumbnail ? {
-                    thumbnail: thumbnail,
-                } : {}),
-            };
-
-            const entryOp = await svc_fsEntry.insert(raw_fsentry);
-
-            await entryOp.awaitDone();
-            resourceService.free(uid);
-
-            const node = await fs.node(new NodeUIDSelector(uid));
-
-            svc_event.emit('fs.create.directory', {
-                node,
-                context: Context.get(),
-            });
-
-            return node;
-        } finally {
-            await lock_handle.unlock();
-        }
+        console.error('This .mkdir should not be called!');
+        throw new Error('This .mkdir should not be called!');
     }
 
     async update_thumbnail ({ context, node, thumbnail }) {
